+++ src/app/(base)/(content)/layout.tsx
@@ -1,0 +1,10 @@
+import { Ticker } from "@/features/base/event/components/ticker";
+
+export default function ContentLayout({ children }: LayoutProps<"/">) {
+  return (
+    <>
+      <Ticker />
+      {children}
+    </>
+  );
+}